refactor: ensure export types are really only used to export types
