+++ lib/plain/entities/team-space-membership.ts
@@ -6,7 +6,10 @@ import type {
   CollectionProp,
   GetOrganizationParams,
 } from '../../common-types'
-import type { TeamSpaceMembershipProps, CreateTeamSpaceMembershipProps } from '../../export-types'
+import type {
+  TeamSpaceMembershipProps,
+  CreateTeamSpaceMembershipProps,
+} from '../../entities/team-space-membership'
 import type { OptionalDefaults } from '../wrappers/wrap'
 
 export type TeamSpaceMembershipPlainClientAPI = {
