+++ lib/plain/entities/usage.ts
@@ -1,5 +1,5 @@
 import type { CollectionProp, QueryParams } from '../../common-types'
-import type { UsageProps } from '../../export-types'
+import type { UsageProps } from '../../entities/usage'
 import type { OptionalDefaults } from '../wrappers/wrap'
 
 export type UsagePlainClientAPI = {
